extract CPU activation kernels into SRP microcrate
