+++ crates/bitnet-cpu-activations/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "bitnet-cpu-activations"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "CPU activation kernels extracted as an SRP microcrate"
+
+[dependencies]
+libm = "0.2.16"
