+++ crates/bitnet-kernels/Cargo.toml
@@ -19,6 +19,7 @@ include = [
 
 [dependencies]
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+bitnet-cpu-activations = { path = "../bitnet-cpu-activations", version = "0.2.1-dev" }
 bitnet-device-probe = { path = "../bitnet-device-probe", version = "0.2.1-dev" }
 thiserror.workspace = true
 bytemuck.workspace = true
@@ -29,7 +30,6 @@ opencl3 = { version = "0.9", optional = true }
 cc = { version = "1.2.46", optional = true }
 bindgen = { version = "0.72.1", optional = true }
 half = { version = "2.7.1", optional = true }
-libm = "0.2.16"
 sysinfo = "0.37.2"
 memory-stats = "1.2.0"
